Fix some whitespace in ZipFile classes
